Align IPC package with lint guidelines
